bblsite-backend/pyproject.toml
[tool.poetry]
name = "bbl"
version = "0.1.0"
description = ""
authors = ["Geol Soo Kim <geolsoo.kim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
Faker = "^8.12.1"
ipython = "^7.27.0"
dj-rest-auth = "^2.1.11"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
